Rename the shortcut editor to avoid repetition
