+++ app/classifier/tasks/shortcut/editor.jsx
