+++ app/pages/lab/workflow.cjsx
@@ -15,7 +15,7 @@ WorkflowCreateForm = require './workflow-create-form'
 workflowActions = require './actions/workflow'
 classnames = require 'classnames'
 
-`import ShortcutEditor from '../../classifier/tasks/shortcut/shortcut-editor';`
+`import ShortcutEditor from '../../classifier/tasks/shortcut/editor';`
 `import FeedbackEditor from './feedback/feedback-editor';`
 
 DEMO_SUBJECT_SET_ID = if process.env.NODE_ENV is 'production'
